inventoryv2

+++ Enemy.h
@@ -9,8 +9,8 @@ public:
 	virtual sf::FloatRect getBoundingBox() = 0;
 	virtual  ~Enemy() {};
 	virtual unsigned int getId() = 0;
-	virtual void setHp(unsigned int damage) = 0;
-	virtual unsigned int getHp() = 0;
+	virtual void setHp(float damage) = 0;
+	virtual float getHp() = 0;
 	virtual unsigned int getDamage() = 0;
 	virtual bool canDealDamage() = 0;
 	virtual void notifyDamageDealt() = 0;
